requirements.txt
django>=4.2.0,<4.3.0
djangorestframework
django-redis==5.3.0
python-decouple==3.8
psycopg2-binary==2.9.6
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers
python-dotenv